# apps/vyve-messenger/docs/openapi.yaml
openapi: 3.0.3
info:
  title: VYVE Messenger API
  version: 1.0.0
  description: |
    API for VYVE Messenger — the private, end-to-end encrypted communication
    layer of LEVI SIOS. Authentication is handled by Cybrus OAuth 2.0.
    
    ## Security Model
    - All endpoints require valid OAuth2 bearer token
    - Message content is end-to-end encrypted client-side
    - Server NEVER receives plaintext message content
    - All timestamps are rounded to 1-hour buckets for metadata privacy

  contact:
    name: VYVE Engineering
    email: engineering@vyve.local

  license:
    name: Proprietary
    url: https://vyve.local/license
servers:
  - url: https://api.vyve.local/v1
    description: Production
  - url: http://localhost:8080/v1
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /conversations:
    get:
      tags: [Conversations]
      summary: List conversations
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: before
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: List of conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Conversation'

    post:
      tags: [Conversations]
      summary: Create a new conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConversationCreate'
      responses:
        '201':
          description: Conversation created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conversation'
  /conversations/{conversation_id}:
    get:
      tags: [Conversations]
      summary: Get conversation details
      parameters:
        - name: conversation_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Conversation details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conversation'

    delete:
      tags: [Conversations]
      summary: Delete a conversation (user-side)
      parameters:
        - name: conversation_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Conversation deleted
  /conversations/{conversation_id}/messages:
    get:
      tags: [Messaging]
      summary: Get messages in a conversation
      parameters:
        - name: conversation_id
          in: path
          required: true
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: before
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: List of encrypted messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EncryptedMessage'

    post:
      tags: [Messaging]
      summary: Send an encrypted message
      description: |
        Server stores the ciphertext but cannot decrypt it.
        Each recipient device receives its own ciphertext variant.
      parameters:
        - name: conversation_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EncryptedMessageUpload'
      responses:
        '201':
          description: Message queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EncryptedMessage'
  /messages/{message_id}:
    get:
      tags: [Messaging]
      summary: Get a specific message
      parameters:
        - name: message_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Encrypted message
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EncryptedMessage'

    delete:
      tags: [Messaging]
      summary: Delete a message (user-side)
      parameters:
        - name: message_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Message deleted
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    User:
      type: object
      properties:
        user_id:
          type: string
          format: uuid
        username:
          type: string
        display_name:
          type: string
        bio:
          type: string
        avatar_url:
          type: string
        public_keys:
          $ref: '#/components/schemas/PublicKeyBundle'
        created_at:
          type: string
          format: date-time
        tier:
          type: string
          enum: [free, creator, pro, developer, business, enterprise]
    PublicKeyBundle:
      type: object
      properties:
        identity_signing_key:
          type: string
          format: byte
          description: Ed25519 public key
        identity_encryption_key:
          type: string
          format: byte
          description: X25519 public key
        key_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
...
    Conversation:
      type: object
      properties:
        conversation_id:
          type: integer
        conversation_type:
          type: string
          enum: [direct, group, channel]
        participants:
          type: array
          items:
            $ref: '#/components/schemas/User'
        created_at:
          type: string
          format: date-time
        last_message:
          $ref: '#/components/schemas/EncryptedMessage'
        unread_count:
          type: integer
        encryption_metadata:
          type: object
          description: Per-participant encryption details
    ConversationCreate:
      type: object
      required: [conversation_type, participants]
      properties:
        conversation_type:
          type: string
          enum: [direct, group, channel]
        participants:
          type: array
          items:
            type: string
            format: uuid
        name:
          type: string
        encryption_keys:
          type: object
          description: Initial encryption keys for the conversation
    EncryptedMessage:
      type: object
      properties:
        message_id:
          type: integer
        conversation_id:
          type: integer
        sender_key_id:
          type: string
          format: uuid
        recipient_key_ids:
          type: array
          items:
            type: string
            format: uuid
        ephemeral_pubkey:
          type: string
          format: byte
        nonce:
          type: string
          format: byte
        ciphertext:
          type: string
          format: byte
        signature:
          type: string
          format: byte
        timestamp:
          type: string
          format: date-time
          description: Rounded to 1-hour bucket
        size_bucket:
          type: string
          enum: [tiny, small, medium, large, xlarge]
        status:
          type: string
          enum: [queued, delivered, read, failed]
    EncryptedMessageUpload:
      type: object
      required: [ciphertext, recipient_key_ids, ephemeral_pubkey, nonce, signature]
      properties:
        recipient_key_ids:
          type: array
          items:
            type: string
            format: uuid
        ephemeral_pubkey:
          type: string
          format: byte
        nonce:
          type: string
          format: byte
        ciphertext:
          type: string
          format: byte
        signature:
          type: string
          format: byte
        attachment_ids:
          type: array
          items:
            type: string
            format: uuid
        reply_to:
          type: integer
          description: ID of message being replied to
        forward_policy:
          type: string
          enum: [allowed, restricted, forbidden]
...